Extract pattern code
