+++ verifier/src/main/java/com/aws/jverify/verifier/JavaToDafnyCompiler.java
@@ -1262,7 +1262,7 @@ public class JavaToDafnyCompiler {
 
     private Expression translateSwitchExpression(JCTree.JCSwitchExpression switchExpr) {
         var origin = toOrigin(switchExpr);
-        var patternBodies = translateSwitchLabels(switchExpr);
+        var patternBodies = new Patterns(this).translateSwitchLabels(switchExpr);
         if (patternBodies == null) {
             return getHole(origin);
         }
@@ -1285,87 +1285,6 @@ public class JavaToDafnyCompiler {
 
         var source = toExpr(switchExpr.getExpression());
         return new NestedMatchExpr(origin, source, translatedCases, true, null);
-    }
-
-    record SwitchLabelPatternAndBody(JCTree.JCCase cas, ExtendedPattern pattern, JCTree body) {}
-
-    /**
-     * Translates the switch labels of the given {@code switch} statement or expression
-     * into the corresponding {@link ExtendedPattern}s,
-     * and returns the patterns along with their corresponding (untranslated) bodies.
-     * Returns {@code null} if an unrecoverable error is reported,
-     * such as if the input tree uses unsupported features.
-     */
-    @Nullable List<SwitchLabelPatternAndBody> translateSwitchLabels(JCTree switchTree) {
-        // JCTree is the first common superclass of JCSwitch and JCSwitchExpression,
-        // so we settle for dynamically checking that the argument is one of them.
-        var cases = switch (switchTree) {
-            case JCTree.JCSwitch switchStmt -> switchStmt.getCases();
-            case JCTree.JCSwitchExpression switchExpr -> switchExpr.getCases();
-            default -> throw new IllegalArgumentException(
-                    "Expected switch statement or expression but got " + switchTree.getClass());
-        };
-
-        // A switch block consists of either *switch rules* (label -> body)
-        // or *switch labeled statement groups* (label: {label:} stmts).
-        // Unlike switch rules, switch labeled statement groups automatically "fall through" without break statements,
-        // but Dafny's match statement/expression can't express that easily.
-        // So for now we only support switch blocks using switch rules.
-        if (cases.getFirst().getCaseKind().equals(JCTree.JCCase.STATEMENT)) {
-            reportError(switchTree, "notSupported", "switch labeled statement group");
-            return null;
-        }
-
-        return cases.stream()
-                .map(cas -> new SwitchLabelPatternAndBody(cas, translateSwitchLabel(cas), cas.getBody()))
-                .toList();
-    }
-
-    /**
-     * Translates the given switch label into a pattern.
-     */
-    private ExtendedPattern translateSwitchLabel(JCTree.JCCase cas) {
-        // Each case has a *switch label*, which is either a *default label* or a *case label*.
-        // A *case label* consists of either:
-        //  - a list of *case constants*
-        //  - a null literal (which the javac AST treats like another case constant)
-        //  - a *case pattern* (not supported)
-
-        // note: if the case label is a null literal, this is the singleton list of the null literal
-        var caseConstants = cas.getExpressions();
-        var defaultLabel = cas.getLabels().stream()
-                .filter(label -> label instanceof JCTree.JCDefaultCaseLabel)
-                .findFirst();
-
-        if (caseConstants.nonEmpty()) {
-            var literals = caseConstants.stream().map(this::translateCaseConstant).toList();
-            return new DisjunctivePattern(toOrigin(cas), false, literals);
-        } else if (defaultLabel.isPresent()) {
-            return makeWildPattern(toOrigin(defaultLabel.get()));
-        } else {
-            reportError(cas, "notSupported", "case pattern");
-            // Return something sensible
-            return makeWildPattern(toOrigin(cas));
-        }
-    }
-
-    /**
-     * Translates the given switch label case constant into a pattern.
-     */
-    private ExtendedPattern translateCaseConstant(JCTree.JCExpression expr) {
-        var origin = toOrigin(expr);
-        final LiteralExpr litExpr;
-        if (expr instanceof JCTree.JCLiteral) {
-            litExpr = (LiteralExpr)toExpr(expr);
-        } else {
-            reportError(expr, "notSupported", "non-literal case constant");
-            litExpr = getHole(origin);
-        }
-        return new LitPattern(origin, false, litExpr);
-    }
-
-    static IdPattern makeWildPattern(IOrigin origin) {
-        return new IdPattern(origin, false, "_", null, null, false);
     }
 
     public @Nullable Type translateType(JCTree tree) {
